pyproject.toml
[project]
name = "agentic-content-scout"
version = "0.1.0"
description = "Personalized content aggregator and recommendation system built with LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0",
    "tavily-python>=0.5",
    "python-dotenv>=1.0",
    "rich>=14.0",
    "prompt_toolkit>=3.0",
]

[project.scripts]
scout = "agentic_content_scout.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
